#142 Refactor the test for CUDA-Q
